Added routing for movies
